Update MQ2Melee - Rearrange infodata
Rearrange infodata by class in pragma regions so they can be collapsed. Purpose - I don't update abilities for every class in alphabetical order of the abilities. I update one class at a time, so I felt it made more sense to be arranged by class. This is it's own commit, so that changes for the classes being updated will be seen in their own commit without all the rearranging diff stuff at the same time.